+++ Classroom/src/main/java/com/Classroom/Classroom/dto/StudentDto.java
@@ -27,7 +27,7 @@ public class StudentDto {
     private Long fatherNumber;
     private Long motherNumber;
     private String bloodGroup;
-    private int cgpa;
+    private double cgpa;
     private int studentConcern;
     private String password;
     private List<StudentSkills> studentSkills;
